update bottom bar nav
