--- app/src/main/java/com/company/khomasi/navigation/AppNavigationDestination.kt
@@ -1,22 +1,0 @@
-package com.company.khomasi.navigation
-
-/**
- * Interface for describing the Now in Android navigation destinations
- */
-
-interface AppNavigationDestination {
-    /**
-     * Defines a specific route this destination belongs to.
-     * Route is a String that defines the path to your composable.
-     * You can think of it as an implicit deep link that leads to a specific destination.
-     * Each destination should have a unique route.
-     */
-    val route: String
-
-    /**
-     * Defines a specific destination ID.
-     * This is needed when using nested graphs via the navigation DLS, to differentiate a specific
-     * destination's route from the route of the entire nested graph it belongs to.
-     */
-    val destination: String
-}
