+++ app/src/main/java/com/company/khomasi/navigation/NavItems.kt
@@ -1,10 +1,12 @@
 package com.company.khomasi.navigation
 
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
 import com.company.khomasi.R
 
 data class NavItem(
-    val label: Int,
-    val icon: Int,
+    @StringRes val label: Int,
+    @DrawableRes val icon: Int,
     val route: String
 )
 
@@ -17,8 +19,8 @@ val listOfNavItems = listOf(
     NavItem(
         label = R.string.my_bookings,
         icon = R.drawable.ticket,
-        route = Screens.MyBookings.name)
-    ,
+        route = Screens.MyBookings.name
+    ),
     NavItem(
         label = R.string.fields,
         icon = R.drawable.soccerball,
