+++ app/src/main/java/com/company/khomasi/presentation/MainActivity.kt
@@ -3,11 +3,11 @@ package com.company.khomasi.presentation
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.material3.Text
-import androidx.compose.ui.res.stringResource
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.company.khomasi.R
-import com.company.khomasi.navigation.AppNavigation
+import com.company.khomasi.presentation.navigation.AppNavigation
+import com.company.khomasi.theme.KhomasiTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -16,10 +16,13 @@ class MainActivity : ComponentActivity() {
         installSplashScreen()
         super.onCreate(savedInstanceState)
         setContent {
-            Text(
-                text = stringResource(id = R.string.app_name)
-            )
-            AppNavigation()
+            KhomasiTheme {
+                Surface(
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    AppNavigation()
+                }
+            }
         }
     }
 }
