+++ app/src/main/java/com/company/khomasi/presentation/navigation/AppNavigation.kt
@@ -1,0 +1,57 @@
+package com.company.khomasi.presentation.navigation
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.company.khomasi.navigation.Screens
+import com.company.khomasi.presentation.ui.screens.FavoriteScreen
+import com.company.khomasi.presentation.ui.screens.HomeScreen
+import com.company.khomasi.presentation.ui.screens.MyBookingsScreen
+import com.company.khomasi.presentation.ui.screens.PlaygroundsScreen
+import com.company.khomasi.theme.KhomasiTheme
+
+@Composable
+fun AppNavigation() {
+    val navController = rememberNavController()
+
+    Scaffold(
+        bottomBar = {
+            BottomNavigationBar(navController = navController)
+        }
+    ) { paddingValues ->
+
+        NavHost(
+            navController = navController,
+            startDestination = Screens.Home.name,
+            modifier = Modifier.padding(paddingValues)
+        ) {
+            composable(route = Screens.Home.name) {
+                HomeScreen()
+            }
+            composable(route = Screens.Favorite.name) {
+                FavoriteScreen()
+            }
+            composable(route = Screens.MyBookings.name) {
+                MyBookingsScreen()
+            }
+            composable(route = Screens.Playgrounds.name) {
+                PlaygroundsScreen()
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Composable
+fun DefaultPreview() {
+    KhomasiTheme {
+        AppNavigation()
+    }
+}
