+++ app/src/main/java/com/company/khomasi/presentation/navigation/BottomNavigationBar.kt
@@ -1,0 +1,49 @@
+package com.company.khomasi.presentation.navigation
+
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.navigation.NavController
+import androidx.navigation.NavDestination.Companion.hierarchy
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.compose.currentBackStackEntryAsState
+import com.company.khomasi.navigation.listOfNavItems
+
+@Composable
+fun BottomNavigationBar(
+    navController: NavController
+) {
+    NavigationBar {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentDestination = navBackStackEntry?.destination
+
+        listOfNavItems.forEach { navItem ->
+            NavigationBarItem(
+                selected = currentDestination?.hierarchy?.any { it.route == navItem.route } == true,
+                onClick = {
+                    navController.navigate(navItem.route) {
+                        popUpTo(navController.graph.findStartDestination().id) {
+                            saveState = true
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                },
+                icon = {
+                    Icon(
+                        painter = painterResource(id = navItem.icon),
+                        contentDescription = null
+                    )
+                },
+                label = {
+                    Text(text = stringResource(id = navItem.label))
+                }
+            )
+        }
+    }
+}
